# ent/ogent/oas_router_gen.go
// Code generated by ogen, DO NOT EDIT.

package ogent

import (
	"net/http"
	"net/url"
	"strings"

	"github.com/ogen-go/ogen/uri"
)

func (s *Server) cutPrefix(path string) (string, bool) {
	prefix := s.cfg.Prefix
	if prefix == "" {
		return path, true
	}
	if !strings.HasPrefix(path, prefix) {
		// Prefix doesn't match.
		return "", false
	}
	// Cut prefix from the path.
	return strings.TrimPrefix(path, prefix), true
}

// ServeHTTP serves http request as defined by OpenAPI v3 specification,
// calling handler that matches the path or returning not found error.
func (s *Server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	elem := r.URL.Path
	elemIsEscaped := false
	if rawPath := r.URL.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
			elemIsEscaped = strings.ContainsRune(elem, '%')
		}
	}

	elem, ok := s.cutPrefix(elem)
	if !ok || len(elem) == 0 {
		s.notFound(w, r)
		return
	}
	args := [1]string{}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/"

			if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'b': // Prefix: "banks"

				if l := len("banks"); len(elem) >= l && elem[0:l] == "banks" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch r.Method {
					case "GET":
						s.handleListBankRequest([0]string{}, elemIsEscaped, w, r)
					case "POST":
						s.handleCreateBankRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, "GET,POST")
					}

					return
				}
				switch elem[0] {
				case '/': // Prefix: "/"

					if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
						elem = elem[l:]
					} else {
						break
					}

					// Param: "id"
					// Match until "/"
					idx := strings.IndexByte(elem, '/')
					if idx < 0 {
						idx = len(elem)
					}
					args[0] = elem[:idx]
					elem = elem[idx:]

					if len(elem) == 0 {
						switch r.Method {
						case "DELETE":
							s.handleDeleteBankRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						case "GET":
							s.handleReadBankRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						case "PATCH":
							s.handleUpdateBankRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, "DELETE,GET,PATCH")
						}

						return
					}
					switch elem[0] {
					case '/': // Prefix: "/"

						if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							break
						}
						switch elem[0] {
						case 'c': // Prefix: "currency-rates"

							if l := len("currency-rates"); len(elem) >= l && elem[0:l] == "currency-rates" {
								elem = elem[l:]
							} else {
								break
							}

							if len(elem) == 0 {
								// Leaf node.
								switch r.Method {
								case "GET":
									s.handleListBankCurrencyRatesRequest([1]string{
										args[0],
									}, elemIsEscaped, w, r)
								default:
									s.notAllowed(w, r, "GET")
								}

								return
							}

						case 'o': // Prefix: "offers"

							if l := len("offers"); len(elem) >= l && elem[0:l] == "offers" {
								elem = elem[l:]
							} else {
								break
							}

							if len(elem) == 0 {
								// Leaf node.
								switch r.Method {
								case "GET":
									s.handleListBankOffersRequest([1]string{
										args[0],
									}, elemIsEscaped, w, r)
								default:
									s.notAllowed(w, r, "GET")
								}

								return
							}

						}

					}

				}

			case 'c': // Prefix: "currency-rates"

				if l := len("currency-rates"); len(elem) >= l && elem[0:l] == "currency-rates" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch r.Method {
					case "GET":
						s.handleListCurrencyRateRequest([0]string{}, elemIsEscaped, w, r)
					case "POST":
						s.handleCreateCurrencyRateRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, "GET,POST")
					}

					return
				}
				switch elem[0] {
				case '/': // Prefix: "/"

					if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
						elem = elem[l:]
					} else {
						break
					}

					// Param: "id"
					// Match until "/"
					idx := strings.IndexByte(elem, '/')
					if idx < 0 {
						idx = len(elem)
					}
					args[0] = elem[:idx]
					elem = elem[idx:]

					if len(elem) == 0 {
						switch r.Method {
						case "DELETE":
							s.handleDeleteCurrencyRateRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						case "GET":
							s.handleReadCurrencyRateRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						case "PATCH":
							s.handleUpdateCurrencyRateRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, "DELETE,GET,PATCH")
						}

						return
					}
					switch elem[0] {
					case '/': // Prefix: "/bank"

						if l := len("/bank"); len(elem) >= l && elem[0:l] == "/bank" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "GET":
								s.handleReadCurrencyRateBankRequest([1]string{
									args[0],
								}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, "GET")
							}

							return
						}

					}

				}

			case 'o': // Prefix: "offers"

				if l := len("offers"); len(elem) >= l && elem[0:l] == "offers" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch r.Method {
					case "GET":
						s.handleListOfferRequest([0]string{}, elemIsEscaped, w, r)
					case "POST":
						s.handleCreateOfferRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, "GET,POST")
					}

					return
				}
				switch elem[0] {
				case '/': // Prefix: "/"

					if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
						elem = elem[l:]
					} else {
						break
					}

					// Param: "id"
					// Match until "/"
					idx := strings.IndexByte(elem, '/')
					if idx < 0 {
						idx = len(elem)
					}
					args[0] = elem[:idx]
					elem = elem[idx:]

					if len(elem) == 0 {
						switch r.Method {
						case "DELETE":
							s.handleDeleteOfferRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						case "GET":
							s.handleReadOfferRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						case "PATCH":
							s.handleUpdateOfferRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, "DELETE,GET,PATCH")
						}

						return
					}
					switch elem[0] {
					case '/': // Prefix: "/bank"

						if l := len("/bank"); len(elem) >= l && elem[0:l] == "/bank" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "GET":
								s.handleReadOfferBankRequest([1]string{
									args[0],
								}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, "GET")
							}

							return
						}

					}

				}

			}

		}
	}
	s.notFound(w, r)
}

// Route is route object.
type Route struct {
	name        string
	summary     string
	operationID string
	pathPattern string
	count       int
	args        [1]string
}

// Name returns ogen operation name.
//
// It is guaranteed to be unique and not empty.
func (r Route) Name() string {
	return r.name
}

// Summary returns OpenAPI summary.
func (r Route) Summary() string {
	return r.summary
}

// OperationID returns OpenAPI operationId.
func (r Route) OperationID() string {
	return r.operationID
}

// PathPattern returns OpenAPI path.
func (r Route) PathPattern() string {
	return r.pathPattern
}

// Args returns parsed arguments.
func (r Route) Args() []string {
	return r.args[:r.count]
}

// FindRoute finds Route for given method and path.
//
// Note: this method does not unescape path or handle reserved characters in path properly. Use FindPath instead.
func (s *Server) FindRoute(method, path string) (Route, bool) {
	return s.FindPath(method, &url.URL{Path: path})
}

// FindPath finds Route for given method and URL.
func (s *Server) FindPath(method string, u *url.URL) (r Route, _ bool) {
	var (
		elem = u.Path
		args = r.args
	)
	if rawPath := u.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
		}
		defer func() {
			for i, arg := range r.args[:r.count] {
				if unescaped, err := url.PathUnescape(arg); err == nil {
					r.args[i] = unescaped
				}
			}
		}()
	}

	elem, ok := s.cutPrefix(elem)
	if !ok {
		return r, false
	}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/"

			if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'b': // Prefix: "banks"

				if l := len("banks"); len(elem) >= l && elem[0:l] == "banks" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch method {
					case "GET":
						r.name = ListBankOperation
						r.summary = "List Banks"
						r.operationID = "listBank"
						r.pathPattern = "/banks"
						r.args = args
						r.count = 0
						return r, true
					case "POST":
						r.name = CreateBankOperation
						r.summary = "Create a new Bank"
						r.operationID = "createBank"
						r.pathPattern = "/banks"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}
				switch elem[0] {
				case '/': // Prefix: "/"

					if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
						elem = elem[l:]
					} else {
						break
					}

					// Param: "id"
					// Match until "/"
					idx := strings.IndexByte(elem, '/')
					if idx < 0 {
						idx = len(elem)
					}
					args[0] = elem[:idx]
					elem = elem[idx:]

					if len(elem) == 0 {
						switch method {
						case "DELETE":
							r.name = DeleteBankOperation
							r.summary = "Deletes a Bank by ID"
							r.operationID = "deleteBank"
							r.pathPattern = "/banks/{id}"
							r.args = args
							r.count = 1
							return r, true
						case "GET":
							r.name = ReadBankOperation
							r.summary = "Find a Bank by ID"
							r.operationID = "readBank"
							r.pathPattern = "/banks/{id}"
							r.args = args
							r.count = 1
							return r, true
						case "PATCH":
							r.name = UpdateBankOperation
							r.summary = "Updates a Bank"
							r.operationID = "updateBank"
							r.pathPattern = "/banks/{id}"
							r.args = args
							r.count = 1
							return r, true
						default:
							return
						}
					}
					switch elem[0] {
					case '/': // Prefix: "/"

						if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							break
						}
						switch elem[0] {
						case 'c': // Prefix: "currency-rates"

							if l := len("currency-rates"); len(elem) >= l && elem[0:l] == "currency-rates" {
								elem = elem[l:]
							} else {
								break
							}

							if len(elem) == 0 {
								// Leaf node.
								switch method {
								case "GET":
									r.name = ListBankCurrencyRatesOperation
									r.summary = "List attached CurrencyRates"
									r.operationID = "listBankCurrencyRates"
									r.pathPattern = "/banks/{id}/currency-rates"
									r.args = args
									r.count = 1
									return r, true
								default:
									return
								}
							}

						case 'o': // Prefix: "offers"

							if l := len("offers"); len(elem) >= l && elem[0:l] == "offers" {
								elem = elem[l:]
							} else {
								break
							}

							if len(elem) == 0 {
								// Leaf node.
								switch method {
								case "GET":
									r.name = ListBankOffersOperation
									r.summary = "List attached Offers"
									r.operationID = "listBankOffers"
									r.pathPattern = "/banks/{id}/offers"
									r.args = args
									r.count = 1
									return r, true
								default:
									return
								}
							}

						}

					}

				}

			case 'c': // Prefix: "currency-rates"

				if l := len("currency-rates"); len(elem) >= l && elem[0:l] == "currency-rates" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch method {
					case "GET":
						r.name = ListCurrencyRateOperation
						r.summary = "List CurrencyRates"
						r.operationID = "listCurrencyRate"
						r.pathPattern = "/currency-rates"
						r.args = args
						r.count = 0
						return r, true
					case "POST":
						r.name = CreateCurrencyRateOperation
						r.summary = "Create a new CurrencyRate"
						r.operationID = "createCurrencyRate"
						r.pathPattern = "/currency-rates"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}
				switch elem[0] {
				case '/': // Prefix: "/"

					if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
						elem = elem[l:]
					} else {
						break
					}

					// Param: "id"
					// Match until "/"
					idx := strings.IndexByte(elem, '/')
					if idx < 0 {
						idx = len(elem)
					}
					args[0] = elem[:idx]
					elem = elem[idx:]

					if len(elem) == 0 {
						switch method {
						case "DELETE":
							r.name = DeleteCurrencyRateOperation
							r.summary = "Deletes a CurrencyRate by ID"
							r.operationID = "deleteCurrencyRate"
							r.pathPattern = "/currency-rates/{id}"
							r.args = args
							r.count = 1
							return r, true
						case "GET":
							r.name = ReadCurrencyRateOperation
							r.summary = "Find a CurrencyRate by ID"
							r.operationID = "readCurrencyRate"
							r.pathPattern = "/currency-rates/{id}"
							r.args = args
							r.count = 1
							return r, true
						case "PATCH":
							r.name = UpdateCurrencyRateOperation
							r.summary = "Updates a CurrencyRate"
							r.operationID = "updateCurrencyRate"
							r.pathPattern = "/currency-rates/{id}"
							r.args = args
							r.count = 1
							return r, true
						default:
							return
						}
					}
					switch elem[0] {
					case '/': // Prefix: "/bank"

						if l := len("/bank"); len(elem) >= l && elem[0:l] == "/bank" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch method {
							case "GET":
								r.name = ReadCurrencyRateBankOperation
								r.summary = "Find the attached Bank"
								r.operationID = "readCurrencyRateBank"
								r.pathPattern = "/currency-rates/{id}/bank"
								r.args = args
								r.count = 1
								return r, true
							default:
								return
							}
						}

					}

				}

			case 'o': // Prefix: "offers"

				if l := len("offers"); len(elem) >= l && elem[0:l] == "offers" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch method {
					case "GET":
						r.name = ListOfferOperation
						r.summary = "List Offers"
						r.operationID = "listOffer"
						r.pathPattern = "/offers"
						r.args = args
						r.count = 0
						return r, true
					case "POST":
						r.name = CreateOfferOperation
						r.summary = "Create a new Offer"
						r.operationID = "createOffer"
						r.pathPattern = "/offers"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}
				switch elem[0] {
				case '/': // Prefix: "/"

					if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
						elem = elem[l:]
					} else {
						break
					}

					// Param: "id"
					// Match until "/"
					idx := strings.IndexByte(elem, '/')
					if idx < 0 {
						idx = len(elem)
					}
					args[0] = elem[:idx]
					elem = elem[idx:]

					if len(elem) == 0 {
						switch method {
						case "DELETE":
							r.name = DeleteOfferOperation
							r.summary = "Deletes a Offer by ID"
							r.operationID = "deleteOffer"
							r.pathPattern = "/offers/{id}"
							r.args = args
							r.count = 1
							return r, true
						case "GET":
							r.name = ReadOfferOperation
							r.summary = "Find a Offer by ID"
							r.operationID = "readOffer"
							r.pathPattern = "/offers/{id}"
							r.args = args
							r.count = 1
							return r, true
						case "PATCH":
							r.name = UpdateOfferOperation
							r.summary = "Updates a Offer"
							r.operationID = "updateOffer"
							r.pathPattern = "/offers/{id}"
							r.args = args
							r.count = 1
							return r, true
						default:
							return
						}
					}
					switch elem[0] {
					case '/': // Prefix: "/bank"

						if l := len("/bank"); len(elem) >= l && elem[0:l] == "/bank" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch method {
							case "GET":
								r.name = ReadOfferBankOperation
								r.summary = "Find the attached Bank"
								r.operationID = "readOfferBank"
								r.pathPattern = "/offers/{id}/bank"
								r.args = args
								r.count = 1
								return r, true
							default:
								return
							}
						}

					}

				}

			}

		}
	}
	return r, false
}