# ent/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Bank Directory API",
    "description": "API для справочника банков",
    "version": "1.0.0"
  },
  "paths": {
    "/banks": {
      "get": {
        "tags": [
          "Bank"
        ],
        "summary": "List Banks",
        "description": "List Banks.",
        "operationId": "listBank",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "what page to render",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "itemsPerPage",
            "in": "query",
            "description": "item count to render per page",
            "schema": {
              "type": "integer",
              "maximum": 255,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "result Bank list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BankList"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "post": {
        "tags": [
          "Bank"
        ],
        "summary": "Create a new Bank",
        "description": "Creates a new Bank and persists it to storage.",
        "operationId": "createBank",
        "requestBody": {
          "description": "Bank to create",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "country": {
                    "type": "string"
                  },
                  "website": {
                    "type": "string"
                  },
                  "logo_url": {
                    "type": "string"
                  },
                  "test": {
                    "type": "string"
                  },
                  "currency_rates": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    }
                  },
                  "offers": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    }
                  }
                },
                "required": [
                  "name",
                  "country"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Bank created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BankCreate"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/banks/{id}": {
      "get": {
        "tags": [
          "Bank"
        ],
        "summary": "Find a Bank by ID",
        "description": "Finds the Bank with the requested ID and returns it.",
        "operationId": "readBank",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the Bank",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bank with requested ID was found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BankRead"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "delete": {
        "tags": [
          "Bank"
        ],
        "summary": "Deletes a Bank by ID",
        "description": "Deletes the Bank with the requested ID.",
        "operationId": "deleteBank",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the Bank",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Bank with requested ID was deleted"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "patch": {
        "tags": [
          "Bank"
        ],
        "summary": "Updates a Bank",
        "description": "Updates a Bank and persists changes to storage.",
        "operationId": "updateBank",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the Bank",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "description": "Bank properties to update",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "country": {
                    "type": "string"
                  },
                  "website": {
                    "type": "string"
                  },
                  "logo_url": {
                    "type": "string"
                  },
                  "test": {
                    "type": "string"
                  },
                  "currency_rates": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    }
                  },
                  "offers": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Bank updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BankUpdate"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/banks/{id}/currency-rates": {
      "get": {
        "tags": [
          "Bank"
        ],
        "summary": "List attached CurrencyRates",
        "description": "List attached CurrencyRates.",
        "operationId": "listBankCurrencyRates",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the Bank",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "what page to render",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "itemsPerPage",
            "in": "query",
            "description": "item count to render per page",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "result Banks list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Bank_CurrencyRatesList"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/banks/{id}/offers": {
      "get": {
        "tags": [
          "Bank"
        ],
        "summary": "List attached Offers",
        "description": "List attached Offers.",
        "operationId": "listBankOffers",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the Bank",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "what page to render",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "itemsPerPage",
            "in": "query",
            "description": "item count to render per page",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "result Banks list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Bank_OffersList"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/currency-rates": {
      "get": {
        "tags": [
          "CurrencyRate"
        ],
        "summary": "List CurrencyRates",
        "description": "List CurrencyRates.",
        "operationId": "listCurrencyRate",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "what page to render",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "itemsPerPage",
            "in": "query",
            "description": "item count to render per page",
            "schema": {
              "type": "integer",
              "maximum": 255,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "result CurrencyRate list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CurrencyRateList"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "post": {
        "tags": [
          "CurrencyRate"
        ],
        "summary": "Create a new CurrencyRate",
        "description": "Creates a new CurrencyRate and persists it to storage.",
        "operationId": "createCurrencyRate",
        "requestBody": {
          "description": "CurrencyRate to create",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "currency": {
                    "type": "string"
                  },
                  "rate": {
                    "type": "number",
                    "format": "double"
                  },
                  "bank": {
                    "type": "integer"
                  }
                },
                "required": [
                  "currency",
                  "rate",
                  "bank"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "CurrencyRate created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrencyRateCreate"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
    "/currency-rates/{id}/bank": {
      "get": {
        "tags": [
          "CurrencyRate"
        ],
        "summary": "Find the attached Bank",
        "description": "Find the attached Bank of the CurrencyRate with the given ID",
        "operationId": "readCurrencyRateBank",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the CurrencyRate",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bank attached to CurrencyRate with requested ID was found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrencyRate_BankRead"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
    "/offers/{id}/bank": {
      "get": {
        "tags": [
          "Offer"
        ],
        "summary": "Find the attached Bank",
        "description": "Find the attached Bank of the Offer with the given ID",
        "operationId": "readOfferBank",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the Offer",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bank attached to Offer with requested ID was found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Offer_BankRead"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "BankList": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "website": {
            "type": "string"
          },
          "logo_url": {
            "type": "string"
          },
          "test": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "country"
        ]
      },
...
    },
    "responses": {
      "400": {
        "description": "invalid input, data invalid",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer"
                },
                "status": {
                  "type": "string"
                },
                "errors": {}
              },
              "required": [
                "code",
                "status"
              ]
            }
          }
        }
      },
...
      "404": {
        "description": "resource not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer"
                },
                "status": {
                  "type": "string"
                },
                "errors": {}
              },
              "required": [
                "code",
                "status"
              ]
            }
          }
        }
      },
...
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080/api/v1",
      "description": "Локальный сервер (по умолчанию)"
    }
  ]
}